+++ market/routes.py
@@ -1,0 +1,13 @@
+from market import app
+from flask import render_template
+from market.models import Item
+# decorator - one step before the function is executed 
+@app.route("/")
+@app.route("/home")
+def home_page():
+    return render_template('home.html')
+
+@app.route("/market")
+def market_page():
+    items = Item.query.all()
+    return render_template('market.html', items = items)
