+++ run.py
@@ -1,0 +1,6 @@
+from market import app
+
+if __name__ == '__main__':
+    app.run(debug=True)
+
+
